refactor(shared): extract apply_inlet_filters_if_enabled into shared.inlet_filters
Plugins that call generate_chat_completion directly need to re-run inlet filters in the same way. Share the helper before expanding its mutation guards.
